Implement Scoreboard Logic & Refactor QuestionAnswerPage
This commit fully implements the scoreboard logic (and by extension, most of the game logic). The QuestionAnswerPage was turned into a JPanel instead of an independent JFrame GUI and initialized in a JOptionPane inside GameplayPage. The results are obtained to increment the scoreboard.

Note: The main game can currently be "played" to completion by clicking the gameboard squares. In the future, we should have this integrated into the dice instead alongside with player pieces on the board.